Fix MacOS build

+++ src/write_to_file.cpp
@@ -5,7 +5,7 @@
 #include <stddef.h>
 
 extern "C" {
-  void WriteToOpenFileC(const char* path, size_t path_len, const char* data, size_t data_len);
+  void WriteToOpenFileC(const char* path, uint64_t path_len, const char* data, uint64_t data_len);
 }
 
 // Avoiding string_view, span...
